// prisma/generated/prisma/models/Lead.ts
export type LeadModel = runtime.Types.Result.DefaultSelection<Prisma.$LeadPayload>
// ...
export type $LeadPayload<ExtArgs extends runtime.Types.Extensions.InternalArgs = runtime.Types.Extensions.DefaultArgs> = {
  name: "Lead"
  objects: {}
  scalars: runtime.Types.Extensions.GetPayloadResult<{
    id: string
    name: string
    email: string
    budgetRange: string
    message: string
    status: $Enums.LeadStatus
    createdAt: Date
    updatedAt: Date
  }, ExtArgs["result"]["lead"]>
  composites: {}
}